Render a diagram spec from makeJoinDiagramSpec as a PNG graphics item.

Requires packages DiagrammeR properly installed to use. Please see vignette('DependencySorting', package = 'replyr') and vignette('joinController', package= 'replyr') for more details.

renderJoinDiagram(diagramSpec, ..., pngFileName = NULL, tempDir = tempdir())

Arguments

diagramSpec
diagram specification from makeJoinDiagramSpec.

...
force later arguments to bind by name.

pngFileName
character, if not null where to write PNG

tempDir
character, if not null tempDir to create/use

Value

DiagrammeR::grViz result as a PNG

Details

This PNG can be smaller than directly including a grViz rendering in markdown. Requires all of DiagrammeR, htmlwidgets, webshot, and magick to be installed with all external dependencies properly installed and configured.

See also

tableDescription, buildJoinPlan, makeJoinDiagramSpec, executeLeftJoinPlan