Bind a list of items by rows (can't use dplyr::bind_rows or dplyr::combine on remote sources). Columns are intersected.

Can't set eagerTempRemoval=TRUE on platforms that don't correctly implement dplyr::compute (for instance Sparklyr prior to full resolution of https://github.com/rstudio/sparklyr/issues/721).

replyr_bind_rows(lst, ..., useDplyrLocal = TRUE, useSparkRbind = TRUE,
  eagerTempRemoval = FALSE,
  tempNameGenerator = makeTempNameGenerator("replyr_bind_rows"))

Arguments

lst
list of items to combine, must be all in same dplyr data service

...
force other arguments to be used by name

useDplyrLocal
logical if TRUE use dplyr for local data.

useSparkRbind
logical if TRUE try to use rbind on Sparklyr data

eagerTempRemoval
logical if TRUE remove temps early.

tempNameGenerator
temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.

Value

single data item

Examples

d <- data.frame(x=1:2)
replyr_bind_rows(list(d,d,d,d,d), useDplyrLocal= FALSE)

#>    x
#> 1  1
#> 2  2
#> 3  1
#> 4  2
#> 5  1
#> 6  2
#> 7  1
#> 8  2
#> 9  1
#> 10 2