Filter a tbl on a column having values in a given set.

Filter a tbl on a column having values in a given set.

replyr_filter(x, cname, values, ..., verbose = TRUE,
  tempNameGenerator = makeTempNameGenerator("replyr_filter"))

Arguments

x
tbl or item that can be coerced into such.

cname
name of the column to test values of.

values
set of values to check set membership of.

...
force later arguments to bind by name.

verbose
logical if TRUE echo warnings

tempNameGenerator
temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.

Value

new tbl with only rows where cname value is in values set.

Examples

values <- c('a','c')
d <- data.frame(x=c('a','a','b','b','c','c'),y=1:6,
                stringsAsFactors=FALSE)
replyr_filter(d,'x',values)

#>   x y
#> 1 a 1
#> 2 a 2
#> 3 c 5
#> 4 c 6