Collect values found in columnsToTakeFrom as tuples (experimental, not fully tested on multiple data suppliers).

Collect values found in columnsToTakeFrom as tuples naming which column the value came from (placed in nameForNewKeyColumn) and value found (placed in nameForNewValueColumn). This is essentially a tidyr::gather, dplyr::melt, or anti-pivot. Similar interface as in the cdata package (though does not perform pre/post condition checks).

replyr_moveValuesToRows(data, nameForNewKeyColumn, nameForNewValueColumn,
  columnsToTakeFrom, ..., na.rm = FALSE, nameForNewClassColumn = NULL,
  tempNameGenerator = makeTempNameGenerator("replyr_moveValuesToRows"))

Arguments

data
data.frame to work with.

nameForNewKeyColumn
character name of column to write new keys in.

nameForNewValueColumn
character name of column to write new values in.

columnsToTakeFrom
character array names of columns to take values from.

...
force later columns to bind by name.

na.rm
logical if TRUE remove rows with NA in nameForNewValueColumn.

nameForNewClassColumn
optional name to land original cell classes to.

tempNameGenerator
temp name generator produced by replyr::makeTempNameGenerator, used to record dplyr::compute() effects.

Value

data item

Examples

d <- data.frame(
  index = c(1, 2, 3),
  info = c('a', 'b', 'c'),
  meas1 = c('m1_1', 'm1_2', 'm1_3'),
  meas2 = c(2.1, 2.2, 2.3),
  stringsAsFactors = FALSE)
replyr_moveValuesToRows(d,
              nameForNewKeyColumn= 'meastype',
              nameForNewValueColumn= 'meas',
              columnsToTakeFrom= c('meas1','meas2'),
              nameForNewClassColumn= 'origMeasurementClass')

#>   index info meastype meas origMeasurementClass
#> 1     1    a    meas1 m1_1            character
#> 2     2    b    meas1 m1_2            character
#> 3     3    c    meas1 m1_3            character
#> 4     1    a    meas2  2.1              numeric
#> 5     2    b    meas2  2.2              numeric
#> 6     3    c    meas2  2.3              numeric

# cdata::moveValuesToRows(d,
#                         nameForNewKeyColumn= 'meastype',
#                         nameForNewValueColumn= 'meas',
#                         columnsToTakeFrom= c('meas1','meas2'),
#                         nameForNewClassColumn= 'origMeasurementClass')