Compute number of rows of a tbl.

Number of row in a table. This function is not "group aware" it returns the total number of rows, not rows per dplyr group. Also replyr_nrow depends on data being returned to count, so some corner cases (such as zero columns) will count as zero rows. In particular work around dplyr issue 2871 https://github.com/tidyverse/dplyr/issues/2871.

replyr_nrow(x)

Arguments

x
tbl or item that can be coerced into such.

Value

number of rows

Examples

d <- data.frame(x=c(1,2))
replyr_nrow(d)

#> [1] 2